refactor: scope packages behind path alias
