+++ apps/web/src/header/Header.tsx
@@ -12,7 +12,7 @@ import {
 } from "@mantine/core";
 import { IconBrandGithub, IconMoonStars, IconSun } from "@tabler/icons-react";
 
-import { useAppStore } from "../store/react-bindings";
+import { useAppStore } from "@/store/react-bindings";
 
 export default function Header() {
   const { events$ } = useAppStore();
